--- src/main/java/com/example/demo/main.java
@@ -1,19 +1,0 @@
-package com.example.demo;
-
-
-import javafx.application.Application;
-import javafx.stage.Stage;
-
-
-public class main extends Application {
-
-    @Override
-    public void start(Stage stage) throws Exception {
-        Unity unity = new Unity();
-        unity.unObjects();
-        unity.RUN(stage);
-    }
-    public static void main(String[] args) {
-        launch();
-    }
-}
