+++ src/main/java/module-info.java
@@ -5,6 +5,8 @@ module com.example.demo {
     requires javafx.fxml;
     requires javafx.graphics;
 
-    opens com.example.demo to javafx.fxml;
-    exports com.example.demo;
+    opens Project.modules.Physics to javafx.fxml;
+    exports Project.modules.Physics;
+    exports Project;
+    opens Project to javafx.fxml;
 }
