+++ src/main/resources/Project/template.fxml
@@ -1,0 +1,84 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.SubScene?>
+<?import javafx.scene.chart.CategoryAxis?>
+<?import javafx.scene.chart.LineChart?>
+<?import javafx.scene.chart.NumberAxis?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.layout.ColumnConstraints?>
+<?import javafx.scene.layout.GridPane?>
+<?import javafx.scene.layout.Region?>
+<?import javafx.scene.layout.RowConstraints?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.paint.Color?>
+<?import javafx.scene.paint.LinearGradient?>
+<?import javafx.scene.paint.Stop?>
+<?import javafx.scene.text.Font?>
+
+<GridPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="800.0" prefWidth="1280.0" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Project.PleaseProvideControllerClassName">
+  <columnConstraints>
+    <ColumnConstraints hgrow="SOMETIMES" maxWidth="1185.6000244140625" minWidth="10.0" prefWidth="1172.79990234375" />
+    <ColumnConstraints hgrow="SOMETIMES" maxWidth="292.0" minWidth="10.0" prefWidth="108.00009765625009" />
+  </columnConstraints>
+  <rowConstraints>
+    <RowConstraints maxHeight="582.4000366210937" minHeight="10.0" prefHeight="443.20000000000005" vgrow="SOMETIMES" />
+    <RowConstraints maxHeight="254.39997558593745" minHeight="10.0" prefHeight="220.79999999999995" vgrow="SOMETIMES" />
+  </rowConstraints>
+   <children>
+      <LineChart fx:id="graph" GridPane.rowIndex="1">
+        <xAxis>
+          <CategoryAxis side="BOTTOM" />
+        </xAxis>
+        <yAxis>
+          <NumberAxis side="LEFT" />
+        </yAxis>
+      </LineChart>
+      <VBox prefHeight="200.0" prefWidth="100.0" spacing="30.0" GridPane.columnIndex="1" GridPane.halignment="CENTER" GridPane.hgrow="NEVER">
+         <children>
+            <Button fx:id="start" mnemonicParsing="false" onAction="#StartClick" style="-fx-background-color: white; -fx-border-color: red;" text="Start" textAlignment="CENTER" textFill="#5c2929" textOverrun="CLIP">
+               <font>
+                  <Font name="Franklin Gothic Heavy" size="13.0" />
+               </font>
+            </Button>
+            <Button fx:id="stop" mnemonicParsing="false" onAction="#StopClick" style="-fx-background-color: white; -fx-border-color: red;" text="Stop" textAlignment="CENTER" textFill="#5c2929" textOverrun="CLIP">
+               <font>
+                  <Font name="Franklin Gothic Heavy" size="13.0" />
+               </font>
+            </Button>
+            <Button fx:id="Reset" mnemonicParsing="false" onAction="#ResetClick" style="-fx-background-color: white; -fx-border-color: red;" text="Reset" textAlignment="CENTER" textFill="#5c2929" textOverrun="CLIP">
+               <font>
+                  <Font name="Franklin Gothic Heavy" size="13.0" />
+               </font>
+            </Button>
+         </children>
+         <GridPane.margin>
+            <Insets />
+         </GridPane.margin>
+         <padding>
+            <Insets left="10.0" top="50.0" />
+         </padding>
+      </VBox>
+      <SubScene fx:id="SubScene" accessibleRole="PARENT" height="551.0" width="1173.0" GridPane.columnSpan="2">
+        <root>
+          <Region />
+        </root>
+         <fill>
+            <LinearGradient endX="1.0" endY="1.0">
+               <stops>
+                  <Stop>
+                     <color>
+                        <Color red="0.6842105388641357" green="0.14191032946109772" blue="0.14191032946109772" />
+                     </color>
+                  </Stop>
+                  <Stop offset="1.0">
+                     <color>
+                        <Color red="0.34736841917037964" green="0.2801247537136078" blue="0.05918128788471222" />
+                     </color>
+                  </Stop>
+               </stops>
+            </LinearGradient>
+         </fill>
+      </SubScene>
+   </children>
+</GridPane>
